pyproject.toml
[tool.poetry]
name = "sai"
version = "0.1.0"
description = "Synthetic A.I."
authors = ["Gianmarco <janmail1990@gmail.com>"]
license = "LICENSE"
readme = "README.md"
packages = [{ include = "sai", from = "src" }]

[build-system]
requires = ["poetry-core>=1.0.0", "cython"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110

[tool.poetry.dependencies]
# Python versions
python = ">=3.10,<3.11"
# Utils
pyrootutils = "*"
loguru = "*"
plac = "^1.3.5"
ipykernel = "*"
jupyter = "*"
jupyter-contrib-nbextensions = "*"
ipywidgets = "*"
# ML
numpy = "*"
catboost = "^1.2"
category-encoders = "^2.6.2"
shap = "^0.42.1"
# Deep Learning
lightning = "^2.0.0"
lightning-bolts = "*"
lightning-transformers = "^0.2.5"
torch = "^2.0.0"
torch-geometric = "^2.3.1"
pytorch-forecasting = { version = "^1.0.0", extras = ["mqf2"] }
opacus = "^1.4.0"
# ML tracking
tensorboard = "^2.14.0"
tensorboardx = "^2.6.2.2"
mlflow = "^2.6.0"
ydata-profiling = "^4.5.1"
dvc = "*"
dvclive = "*"
rich = "^13.5.2"
# Configs
hydra-core = "*"
# UI
gradio = "^3.45.2"
# MusicLM
fairseq = { file = "wheels/fairseq-0.12.2-cp310-cp310-macosx_14_0_arm64.whl", platform = "darwin" }
musiclm-pytorch = "^0.2.8"
datasets = { extras = ["audio"], version = "^2.14.5" }
yt-dlp = "^2023.9.24"

[tool.poetry.dev-dependencies]
cython = "^3.0.2"
setuptools = "*"
black = { version = "*", extras = ["jupyter"] }
ipython = "*"
isort = "*"
mypy = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-pylint = "*"
pytest-mypy = "*"
pytest-testmon = "*"
pytest-xdist = "*"
virtualenv = "*"
nbmake = "*"
